+++ selfdrive/car/hyundai/fingerprints.py
@@ -1593,6 +1593,14 @@ FW_VERSIONS = {
       b'\xf1\x00NQ5__               1.01 1.03 99110-CH000         ',
     ],
   },
+  CAR.KIA_SPORTAGE_PHEV_5TH_GEN: {
+    (Ecu.fwdRadar, 0x7d0, None): [
+      b'\xf1\x00NQ5__               1.01 1.03 99110-CH000         ',
+    ],
+    (Ecu.fwdCamera, 0x7c4, None): [
+      b'\xf1\x00NQ5 FR_CMR AT USA LHD 1.00 1.01 99211-P1060 680',
+    ],
+  },
   CAR.SANTA_CRUZ_1ST_GEN: {
     (Ecu.fwdCamera, 0x7c4, None): [
       b'\xf1\x00NX4 FR_CMR AT USA LHD 1.00 1.00 99211-CW000 14M',